Moved things to devS7plc.h
