+++ devS7plc.c
@@ -4,14 +4,7 @@
 #include <stdio.h>
 #include <limits.h>
 
-#include <alarm.h>
-#include <dbAccess.h>
-#include <recGbl.h>
-#include <devSup.h>
 #include <devLib.h>
-#include <errlog.h>
-#include <menuFtype.h>
-
 #include <epicsVersion.h>
 
 #include <biRecord.h>
@@ -41,48 +34,8 @@
 #endif
 
 #include "drvS7plc.h"
+#include "devS7plc.h"
 
-#define isnan(x) ((x)!=(x))
-
-/* suppress compiler warning concerning long long with __extension__ */
-#if (!defined __GNUC__) || (__GNUC__ < 2) || (__GNUC__ == 2 && __GNUC_MINOR__ < 8)
-#define __extension__
-#endif
-
-#ifndef epicsUInt64
-#if (LONG_MAX > 2147483647L)
-#define epicsUInt64 unsigned long
-#define CONV64 "%016lx"
-#else
-#define epicsUInt64 unsigned long long
-#define CONV64 "%016llx"
-#endif
-#endif
-
-#define S7MEM_TIME 100
-
-typedef struct {              /* Private structure to save IO arguments */
-    s7plcStation *station;    /* Card id */
-    unsigned short offs;      /* Offset (in bytes) within memory block */
-    unsigned short bit;       /* Bit number (0-15) for bi/bo */
-    unsigned short dtype;     /* Data type */
-    unsigned short dlen;      /* Data length (in bytes) */
-    epicsInt32 hwLow;         /* Hardware Low limit */
-    epicsInt32 hwHigh;        /* Hardware High limit */
-} S7memPrivate_t;
-
-STATIC int s7plcIoParse(char* recordName, char *parameters, S7memPrivate_t *);
-STATIC long s7plcGetInIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt);
-STATIC long s7plcGetOutIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt);
-
-struct devsup {
-    long      number;
-    DEVSUPFUN report;
-    DEVSUPFUN init;
-    DEVSUPFUN init_record;
-    DEVSUPFUN get_ioint_info;
-    DEVSUPFUN io;
-};
 
 /* stringout for address ********************************************/
 
@@ -423,7 +376,7 @@ epicsExportAddress(dset, s7plcCalcout);
 
 /*********  Support for "I/O Intr" for input records ******************/
 
-STATIC long s7plcGetInIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt)
+long s7plcGetInIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt)
 {
     S7memPrivate_t* p = record->dpvt;
     if (p == NULL)
@@ -438,7 +391,7 @@ STATIC long s7plcGetInIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt)
 
 /*********  Support for "I/O Intr" for output records ****************/
 
-STATIC long s7plcGetOutIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt)
+long s7plcGetOutIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt)
 {
     S7memPrivate_t* p = record->dpvt;
     if (p == NULL)
@@ -470,7 +423,7 @@ STATIC long s7plcGetOutIntInfo(int cmd, dbCommon *record, IOSCANPVT *ppvt)
  *          <hwHigh>    - raw value that mapps to EGUF
  **********************************************************************/
 
-STATIC int s7plcIoParse(char* recordName, char *par, S7memPrivate_t *priv)
+int s7plcIoParse(char* recordName, char *par, S7memPrivate_t *priv)
 {
     char devName[255];
     char *p = par, separator;
